requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
Pillow==10.2.0
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
groq>=0.37.0
requests==2.31.0
